sequencer::api

Module data_source

Structs

PublicHotShotConfig

This struct defines the public Hotshot configuration parameters. Our config module features a GET endpoint accessible via the route /hotshot to display the hotshot config parameters. Hotshot config has sensitive information like private keys and such fields are excluded from this struct.

PublicNetworkConfig

PublicValidatorConfig

This struct defines the public Hotshot validator configuration. Private key and state key pairs are excluded for security reasons.

Traits

CatchupDataSource 🔒

DataSourceOptions

HotShotConfigDataSource 🔒

NodeStateDataSource 🔒

SequencerDataSource

A data source with sequencer-specific functionality.

StakeTableDataSource 🔒

StateSignatureDataSource 🔒

SubmitDataSource 🔒

Functions

provider

Create a provider for fetching missing data from a list of peer query services.

Type Aliases

Provider

Provider for fetching missing data for the query service.