sequencer::api::data_source
Trait CatchupDataSource

pub(crate) trait CatchupDataSource: Sync {
    // Required methods
    fn get_accounts(
        &self,
        instance: &NodeState,
        height: u64,
        view: ViewNumber,
        accounts: &[FeeAccount],
    ) -> impl Send + Future<Output = Result<FeeMerkleTree>>;
    fn get_frontier(
        &self,
        instance: &NodeState,
        height: u64,
        view: ViewNumber,
    ) -> impl Send + Future<Output = Result<BlocksFrontier>>;
    fn get_chain_config(
        &self,
        commitment: Commitment<ChainConfig>,
    ) -> impl Send + Future<Output = Result<ChainConfig>>;

    // Provided method
    fn get_account(
        &self,
        instance: &NodeState,
        height: u64,
        view: ViewNumber,
        account: FeeAccount,
    ) -> impl Send + Future<Output = Result<AccountQueryData>> { ... }
}

Required Methods

fn get_accounts( &self, instance: &NodeState, height: u64, view: ViewNumber, accounts: &[FeeAccount], ) -> impl Send + Future<Output = Result<FeeMerkleTree>>

Get the state of the requested accounts.

The state is fetched from a snapshot at the given height and view, which must correspond! height is provided to simplify lookups for backends where data is not indexed by view. This function is intended to be used for catchup, so view should be no older than the last decided view.

fn get_frontier( &self, instance: &NodeState, height: u64, view: ViewNumber, ) -> impl Send + Future<Output = Result<BlocksFrontier>>

Get the blocks Merkle tree frontier.

The state is fetched from a snapshot at the given height and view, which must correspond! height is provided to simplify lookups for backends where data is not indexed by view. This function is intended to be used for catchup, so view should be no older than the last decided view.

fn get_chain_config( &self, commitment: Commitment<ChainConfig>, ) -> impl Send + Future<Output = Result<ChainConfig>>

Provided Methods

fn get_account( &self, instance: &NodeState, height: u64, view: ViewNumber, account: FeeAccount, ) -> impl Send + Future<Output = Result<AccountQueryData>>

Get the state of the requested account.

The state is fetched from a snapshot at the given height and view, which must correspond! height is provided to simplify lookups for backends where data is not indexed by view. This function is intended to be used for catchup, so view should be no older than the last decided view.

Object Safety

This trait is not object safe.