sequencer::api

Module sql

Functions

fee_header_dependencies 🔒

Get the dependencies needed to apply the STF to the given list of headers.

get_leaf_from_proposal 🔒

load_accounts 🔒

load_chain_config 🔒

load_frontier 🔒

load_reward_accounts 🔒

reconstruct_state 🔒

Reconstructs the ValidatedState from a specific block height up to a given view.

reward_header_dependencies 🔒

Identifies all reward accounts required to replay the State Transition Function for the given leaf proposals. These accounts should be present in the Merkle tree before applying the STF to avoid recursive catchup (i.e., STF triggering another catchup).

Type Aliases

DataSource