sequencer::api::test_helpers

Function catchup_test_helper

pub async fn catchup_test_helper(opt: impl FnOnce(Options) -> Options)

Test the catchup API with custom options.

The opt function can be used to modify the Options which are used to start the server. By default, the options are the minimal required to run this test (configuring a port and enabling the catchup API). opt may add additional functionality (e.g. adding a query module to test a different initialization path) but should not remove or modify the existing functionality (e.g. removing the catchup module or changing the port).