Module sequencer::block::full_payload::ns_table

Types related to a namespace table.

All code that needs to know the binary format of a namespace table is restricted to this file.

See NsTable for a full specification of the binary format of a namespace table.

Structs

NsIndex

Index for an entry in a ns table.

NsIter 🔒

Return type for [Payload::ns_iter].

NsTable

Raw binary data for a namespace table.

NsTableBuilder

Constants

NS_ID_BYTE_LEN 🔒

NS_OFFSET_BYTE_LEN 🔒

NUM_NSS_BYTE_LEN 🔒

Byte lengths for the different items that could appear in a namespace table.