sequencer::catchup

Function local_and_remote

pub(crate) async fn local_and_remote(
    persistence: impl SequencerPersistence,
    remote: impl StateCatchup + 'static,
) -> Arc<dyn StateCatchup>

A catchup implementation that falls back to a remote provider, but prefers a local provider when supported.