Module sequencer::hotshot_commitment

Structs

CommitmentTaskOptions

Enums

SyncError 🔒

Error type during synchronization between data sources (e.g. L1, query services)

Constants

RETRY_DELAY 🔒

Traits

HotShotDataSource 🔒

Trait for generalized query service for parts of the HotShot data involved in syncing with L1 contract

Functions

build_sequence_batches_txn 🔒

prepare the transaction from new leaves (with QC) from HotShot

run_hotshot_commitment_task

main logic for the commitment task, which sync the latest blocks from HotShot to L1 contracts

sequence 🔒

sync_with_l1 🔒

main logic for catching up with HotShot contract on L1

Type Aliases

HotShotClient 🔒