Module sequencer::l1_client

L1 Client

L1Client defines the interface that Espresso Sequencer nodes use to interact with the L1. Sequencer nodes must read from the L1 to populate Espresso block metadata such as the L1 chain height, which is used to facilitate bridging between the L1 and any rollups which are running on the sequencer.

This client runs asynchronously, updating an in-memory snapshot of the relevant L1 information each time a new L1 block is published. This design as a few advantages:

The L1 client is not synchronized with or triggered by HotShot consensus. It can run in pace with the L1, which makes it easy to use a subscription instead of polling for new blocks, vastly reducing the number of L1 RPC calls we make.

HotShot block building does not interact directly with the L1; it simply reads the latest snapshot from the client’s memory. This means that block production is instant and infallible. Any failures or delays in interacting with the L1 will just slow the updating of the L1 snapshot, which will cause the block builder to propose with a slightly old snapshot, but they will still be able to propose on time.

Structs

L1BlockInfo

L1Client

An Http Provider and configuration to interact with the L1.

L1Snapshot

Functions

get_finalized_block 🔒