Module sequencer::network::cdn

Structs

BrokerDef

The broker definition for the Push CDN. Uses the TCP protocol and trusted middleware.

ClientDef

The client definition for the Push CDN. Uses the Quic protocol and no middleware. Differs from the user definition in that is on the client-side.

ProductionDef

The production run definition for the Push CDN. Uses the real protocols and a Redis discovery client.

TestingDef

The testing run definition for the Push CDN. Uses the real protocols, but with an embedded discovery client.

UserDef

The user definition for the Push CDN. Uses the Quic protocol and untrusted middleware.

WrappedSignatureKey

A wrapped SignatureKey. We need to implement the Push CDN’s SignatureScheme trait in order to sign and verify messages to/from the CDN.

Enums

Topic

The enum for the topics we can subscribe to in the Push CDN