sequencer::persistence::fs

Function epoch_files

fn epoch_files(
    dir: impl AsRef<Path>,
) -> Result<impl Iterator<Item = (EpochNumber, PathBuf)>>

Get all paths under dir whose name is of the form .txt. Should probably be made generic and merged with view_files.