sequencer

Module state_signature

Utilities for generating and storing the most recent light client state signatures.

Modules

relay_server

A relay server that’s collecting and serving the light client state signatures

Structs

StateSignatureMemStorage

A rolling in-memory storage for the most recent light client state signatures.

StateSigner

Constants

SIGNATURE_STORAGE_CAPACITY 🔒

Capacity for the in memory signature storage.

Functions

form_light_client_state 🔒

hash_bytes_to_field 🔒

static_stake_table_commitment

Helper function for stake table commitment

Type Aliases

StakeTableCommitmentType

Type for stake table commitment