sequencer::state_signature

Module relay_server

A relay server that’s collecting and serving the light client state signatures

Structs

Options

configurability options for the web server

StateRelayServerState 🔒

State that checks the light client state update and the signature collection

Traits

StateRelayServerDataSource

Functions

define_api 🔒

Set up APIs for relay server

run_relay_server

Type Aliases

Error 🔒

State 🔒