sequencer

Module testing

Structs

LegacyBuilderImplementation 🔒

TestConfig

TestConfigBuilder

Constants

BUILDER_CHANNEL_CAPACITY_FOR_TEST 🔒

STAKE_TABLE_CAPACITY_FOR_TEST 🔒

Functions

run_legacy_builder

run_test_builder

staking_priv_keys

wait_for_decide_on_handle

wait_for_epochs

Waits until a node has reached the given target epoch (exclusive). The function returns once the first event indicates an epoch higher than target_epoch.

Type Aliases

AnvilFillProvider 🔒