Crate sequencer

Re-exports

pub use block::payload::Payload;

pub use genesis::Genesis;

pub use l1_client::L1BlockInfo;

pub use options::Options;

pub use state::ValidatedState;

pub use transaction::NamespaceId;

pub use transaction::Transaction;

Modules

api

block

catchup

chain_config 🔒

context

eth_signature_key

genesis

header 🔒

hotshot_commitment

l1_client

L1 Client

network

options

persistence

Sequencer node persistence.

state

state_signature

Utilities for generating and storing the most recent light client state signatures.

testing

transaction

Structs

BlockBuildingSnafu 🔒

SNAFU context selector for the Error::BlockBuilding variant

ChainConfig

Global variables for an Espresso blockchain.

GenesisWrongSizeSnafu 🔒

SNAFU context selector for the Error::GenesisWrongSize variant

Header

A header is like a [Block] with the body replaced by a digest.

IncorrectParentSnafu 🔒

SNAFU context selector for the Error::IncorrectParent variant

IncorrectViewSnafu 🔒

SNAFU context selector for the Error::IncorrectView variant

L1Params

MerkleTreeSnafu 🔒

SNAFU context selector for the Error::MerkleTreeError variant

MissingGenesisSnafu 🔒

SNAFU context selector for the Error::MissingGenesis variant

NetworkParams

Node

The Sequencer node is generic over the hotshot CommChannel.

NodeState

SeqTypes

UnexpectedGenesisSnafu 🔒

SNAFU context selector for the Error::UnexpectedGenesis variant

Enums

Error

Functions

empty_builder_commitment

init_node

Type Aliases

Event

Leaf

PrivKey

PubKey