Crate sequencer

Re-exports

pub use genesis::Genesis;

pub use options::Options;

Modules

api

catchup

context

external_event_handler 🔒

Should probably rename this to “external” or something

genesis

network

options

persistence

Sequencer node persistence.

proposal_fetcher 🔒

request_response 🔒

run 🔒

state

state_signature

Utilities for generating and storing the most recent light client state signatures.

Structs

L1Params

NetworkParams

Node

The Sequencer node is generic over the hotshot CommChannel.

Constants

RECENT_STAKE_TABLES_LIMIT

Functions

empty_builder_commitment

init_node

main

Type Aliases

SequencerApiVersion