Trait sequencer::api::data_source::CatchupDataSource

pub(crate) trait CatchupDataSource {
    // Provided methods
    fn get_account(
        &self,
        _height: u64,
        _view: ViewNumber,
        _account: Address,
    ) -> impl Send + Future<Output = Result<AccountQueryData>> { ... }
    fn get_frontier(
        &self,
        _height: u64,
        _view: ViewNumber,
    ) -> impl Send + Future<Output = Result<BlocksFrontier>> { ... }
    fn get_chain_config(
        &self,
        _commitment: Commitment<ChainConfig>,
    ) -> impl Send + Future<Output = Result<ChainConfig>> { ... }
}

Provided Methods

fn get_account( &self, _height: u64, _view: ViewNumber, _account: Address, ) -> impl Send + Future<Output = Result<AccountQueryData>>

Get the state of the requested account.

The state is fetched from a snapshot at the given height and view, which must correspond! height is provided to simplify lookups for backends where data is not indexed by view. This function is intended to be used for catchup, so view should be no older than the last decided view.

fn get_frontier( &self, _height: u64, _view: ViewNumber, ) -> impl Send + Future<Output = Result<BlocksFrontier>>

Get the blocks Merkle tree frontier.

The state is fetched from a snapshot at the given height and view, which must correspond! height is provided to simplify lookups for backends where data is not indexed by view. This function is intended to be used for catchup, so view should be no older than the last decided view.

fn get_chain_config( &self, _commitment: Commitment<ChainConfig>, ) -> impl Send + Future<Output = Result<ChainConfig>>

Object Safety

This trait is not object safe.